:nail_care: Update css files for send LSK to properly aligne buttons
