+++ i18n/locales/en/common.json
@@ -254,7 +254,6 @@
   "Rank": "Rank",
   "Rank / Status": "Rank / Status",
   "Read More": "Read More",
-  "Receiver": "Receiver",
   "Recent searches": "Recent searches",
   "Recipient": "Recipient",
   "Redo": "Redo",
@@ -271,7 +270,6 @@
   "Request": "Request",
   "Request LSK": "Request LSK",
   "Request specific amount": "Request specific amount",
-  "Request specific amount (LSK)": "Request specific amount (LSK)",
   "Required": "Required",
   "Reset": "Reset",
   "Reset timer & continue": "Reset timer & continue",
@@ -318,7 +316,6 @@
   "Sign back in": "Sign back in",
   "Sign in": "Sign in",
   "Sorry": "Sorry",
-  "Specify Amount": "Specify Amount",
   "Standby": "Standby",
   "Start here": "Start here",
   "Start the tour": "Start the tour",
@@ -357,7 +354,6 @@
   "Transaction is being processed and will be confirmed. It may take up to 15 minutes to be secured in the blockchain.": "Transaction is being processed and will be confirmed. It may take up to 15 minutes to be secured in the blockchain.",
   "Transaction not found": "Transaction not found",
   "Transactions can’t be reversed": "Transactions can’t be reversed",
-  "Transfer": "Transfer",
   "Try Again": "Try Again",
   "Try again": "Try again",
   "Try using menu for navigation.": "Try using menu for navigation.",
@@ -424,7 +420,6 @@
   "Your passphrase is used to access your Lisk ID.": "Your passphrase is used to access your Lisk ID.",
   "Your registration is secured on the blockchain": "Your registration is secured on the blockchain",
   "Your registration is secured on the blockchain.": "Your registration is secured on the blockchain.",
-  "Your request": "Your request",
   "Your second passphrase registration is being processed and will be confirmed.": "Your second passphrase registration is being processed and will be confirmed.",
   "Your selection": "Your selection",
   "Your session was timed out after 10 minutes of no network activity. You may continue to use certain sections of your Lisk Hub or sign back in to access everything.": "Your session was timed out after 10 minutes of no network activity. You may continue to use certain sections of your Lisk Hub or sign back in to access everything.",
