--- src/components/receive/request.css
@@ -1,129 +1,0 @@
-@import '../app/variables.css';
-
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-
-  & header {
-    padding: 20px 0px;
-    line-height: 36px;
-
-    & .mobileTitle {
-      display: none;
-    }
-
-    & .desktopTitle {
-      display: block;
-    }
-
-    & .subTitle {
-      font-size: var(--subtitle-font-size);
-    }
-
-    & h2 {
-      margin-bottom: 0;
-    }
-
-    & h3 {
-      display: inline-block;
-      margin: 0;
-    }
-  }
-
-  & .body {
-    text-align: center;
-    margin: 2vh 0px; /* stylelint-disable-line */
-  }
-
-  & .emailLink {
-    text-decoration: none;
-    color: var(--color-primary-standard);
-    font-weight: var(--font-weight-bold);
-    font-size: 14px;
-  }
-
-  & .copy {
-    color: var(--color-grayscale-medium);
-    display: block;
-    margin-bottom: 15px;
-    margin-left: -5px;
-    font-size: 14px;
-    font-weight: var(--font-weight-bold);
-    word-wrap: break-word;
-  }
-
-  & .qrCode {
-    margin-bottom: 20px;
-    cursor: pointer;
-
-    &.magnified {
-      & canvas {
-        max-width: 170px;
-        width: 100% !important;
-        height: auto !important;
-      }
-    }
-
-    &.minimized {
-      & canvas {
-        width: 60px !important;
-        height: auto !important;
-      }
-    }
-  }
-
-  & .backButton {
-    min-width: 0;
-    width: 100%;
-  }
-}
-
-.paddingLeft {
-  padding-left: 10px;
-}
-
-.accountVisual {
-  position: absolute;
-  left: -40px;
-  top: -15px;
-}
-
-.disabledInput input:disabled {
-  opacity: 1;
-  border: none;
-}
-
-@media (--small-viewport) {
-  .wrapper {
-    & .subTitle {
-      display: none;
-    }
-
-    & .copy {
-      font-size: 12px;
-    }
-  }
-}
-
-@media (--medium-viewport) {
-  .wrapper {
-    & header {
-      & h2 {
-        line-height: initial;
-        margin: 0;
-      }
-
-      & .mobileTitle {
-        display: block;
-      }
-
-      & .desktopTitle {
-        display: none;
-      }
-    }
-  }
-}
