--- src/components/request/input.css
@@ -1,7 +1,0 @@
-.input:first-of-type .label {
-  padding-left: 60px;
-}
-
-.input input:first-of-type {
-  padding-left: 60px;
-}
