+++ src/components/request/receive.css
