+++ src/components/request/request.test.js
@@ -1,28 +1,30 @@
 import React from 'react';
 import { expect } from 'chai';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import PropTypes from 'prop-types';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import { spy } from 'sinon';
 import i18n from '../../i18n';
-import ReceiveHowItWorks from './receiveHowItWorks';
+import Request from './request';
 
-
-describe('ReceiveHowItWorks', () => {
+describe('Render Request', () => {
   let wrapper;
   let props;
   let store;
 
   beforeEach(() => {
+    const history = {
+      location: {
+        pathname: 'request',
+        search: '',
+      },
+      push: () => {},
+    };
+
     props = {
+      history,
       address: '12345L',
       t: key => key,
-      nextStep: spy(),
-      prevStep: spy(),
-      settingsUpdated: spy(),
-      isRequestHowItWorksDisable: false,
-      status: 'foward',
     };
 
     store = configureMockStore([thunk])({
@@ -36,7 +38,7 @@ describe('ReceiveHowItWorks', () => {
       },
     });
 
-    wrapper = shallow(<ReceiveHowItWorks {...props}/>, {
+    wrapper = mount(<Request {...props}/>, {
       context: { store, i18n },
       childContextTypes: {
         store: PropTypes.object.isRequired,
@@ -45,28 +47,24 @@ describe('ReceiveHowItWorks', () => {
     });
   });
 
-  it('render ReceiveHowItWorks component', () => {
+  it('render Request component', () => {
     expect(wrapper.exists()).to.equal(true);
   });
 
-  it('continue to next component after click on X icon', () => {
-    wrapper.find('.okay-button').simulate('click');
-    wrapper.update();
-    expect(props.nextStep).to.have.been.calledWith();
-  });
-
-  it('display message about how it works', () => {
-    props.isRequestHowItWorksDisable = true;
-    wrapper = shallow(<ReceiveHowItWorks {...props}/>, {
+  it('render Request component without address', () => {
+    props.address = undefined;
+    wrapper = mount(<Request {...props}/>, {
       context: { store, i18n },
       childContextTypes: {
         store: PropTypes.object.isRequired,
         i18n: PropTypes.object.isRequired,
       },
     });
+    expect(wrapper.exists()).to.equal(true);
+  });
 
-    wrapper.find('.okay-button').simulate('click');
+  it('go to transaction page on click back button', () => {
+    wrapper.find('Button').at(0).simulate('click');
     wrapper.update();
-    expect(props.nextStep).to.have.been.calledWith();
   });
 });
