+++ src/components/request/steps/confirmation/requestForm.css
@@ -1,4 +1,4 @@
-@import '../app/variables.css';
+@import '../../../app/variables.css';
 
 .requestForm {
   display: flex;
