+++ src/components/request/steps/confirmation/requestForm.test.js
@@ -5,7 +5,7 @@ import { mount } from 'enzyme';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 
-import RequestForm from './receiveForm';
+import RequestForm from './requestForm';
 
 describe('RequestForm', () => {
   let wrapper;
