+++ src/components/resultBox/resultBox.css
@@ -17,10 +17,15 @@
 .resultBox {
   text-align: center;
   justify-content: space-evenly;
+  align-items: center;
   flex-flow: column;
   display: flex;
   width: 100%;
   height: 100%;
+
+  & > div {
+    margin-bottom: 100px;
+  }
 
   & .header {
     position: relative;
@@ -28,17 +33,6 @@
 
   & header > h2 {
     font-size: 28px;
-  }
-
-  & footer {
-    padding: 0px;
-    display: flex;
-    justify-content: center;
-
-    & > button {
-      width: calc(100% / 2 - 20px);
-      margin: 0;
-    }
   }
 
   & p {
@@ -74,7 +68,23 @@
     width: 100%;
   }
 
-  & .okButton {
-    margin-left: 20px;
+  & footer {
+    padding: 0px;
+    display: flex;
+    justify-content: center;
+    position: absolute;
+    bottom: 58px;
+    width: 50%;
+
+    & > div > button {
+      width: 160px !important;
+      min-width: 160px !important;
+      margin: 0;
+    }
+
+    & .okButton {
+      width: 160px !important;
+      min-width: 160px !important;
+    }
   }
 }
